Refactor ZipMaker in preparation of it being used to create the zip for the GwtZipFixer.
